==> code/00_init.R <==
# ------------------------------------------------------------------#
# Only do this if you're installing this repo with a clean start!!  #
# ------------------------------------------------------------------#


# We have logged all packages used, and these will be installed,
# when you run this:
renv::init() # select 1 "Restore the project from the lockfile."
#  select yes (y)
# This may take some time to install. Make a tea.
